Update Jenkins
Updated the jenkins tutorial to match the prometheus tutorial. More investigation and work needs to happen on cert manager as the TLS cert is not working for the specified DNS name. Instead it shows as created for ingress.local. Need to investigate Let's Encrypt as an option here instead of a standard CA.

Signed-off-by: Kendrick Coleman <kendrickc@vmware.com>